requirements.txt
Flask==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.1.1
tiktoken==0.5.2
bcrypt==4.1.2
schedule==1.2.0
flask-cors==6.0.1
gunicorn==23.0.0
pandas==2.3.0
openpyxl==3.1.5
typing-extensions==4.14.1
xlrd==2.0.2
Unidecode==1.3.7
rapidfuzz==3.6.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
Pillow==10.4.0